style : 프로토콜 extension을 통해서 채택
