+++ CodeStarterCamp_Week5/PersonModel/Person.swift
@@ -1,0 +1,19 @@
+//
+//  Person.swift
+//  CodeStarterCamp_Week5
+//
+//  Created by 이정민 on 2022/05/10.
+//
+
+import Foundation
+
+class Person {
+    let name: String
+    let height: Int
+
+    init(name: String, height: Int) {
+        self.name = name
+        self.height = height
+    }
+}
+
